+++ spec/dummy/client/entries/app-kitten.js
@@ -237,7 +237,8 @@ import {
 import { TabBar } from 'kitten/components/navigation/tab-bar'
 import {
   KarlHorizontalNav,
-  KarlBrowsingHorizontalNav,
+  KarlTouchHorizontalNav,
+  KarlTouchHorizontalNavWithNotification,
 } from 'kitten/karl/navigation/horizontal-nav'
 import { KarlPaginationExamples } from 'kitten/karl/navigation/pagination'
 
@@ -512,7 +513,8 @@ ReactOnRails.register({
   // Navigation
   TabBar,
   KarlHorizontalNav,
-  KarlBrowsingHorizontalNav,
+  KarlTouchHorizontalNav,
+  KarlTouchHorizontalNavWithNotification,
   KarlPaginationExamples,
 
   // Notifications